+++ R/blocked_anticlustering.R
@@ -1,0 +1,68 @@
+
+# Use blocking for anticlustering (anticluster within a block, which a level of a cagegorical variable, 
+# or a stratum). When working on subsequent blocks, it "remembers" the assignment of prior blocks
+# when optimizing an anticlustering objective.
+
+# including constraints (must-link /cannot-link) would require re-indexing, I will at first not allow this
+# because it is additional work for me.
+
+blocked_anticlustering <- function(
+    x, K, objective = "diversity", method = "exchange", preclustering = FALSE, 
+    categories = NULL, repetitions = NULL, standardize = FALSE, blocks
+) {
+
+  validate_input(K, "K", len = 1) # here we can only generate equal-sized groups
+  
+  N <- nrow(x)
+
+  if (!argument_exists(blocks)) stop("Can only use blocked anticlustering when argument 'blocks' is given.")  
+  blocks <- merge_into_one_variable(blocks) # use "strata"
+  
+  n_blocks <- length(unique(blocks))
+  blocksizes <- table(blocks)
+
+  # Anticlustering with blocking: 
+  condition_blocked <- rep(NA, N)
+  for (i in 1:n_blocks) {
+    previous_groups <- condition_blocked # just for asserting at the end that no previous assignments were changed
+    select <- blocks <= i
+    input <- x[select, , drop = FALSE] 
+    target_groups <- table(initialize_clusters(N = sum(select), K = K, NULL))
+    initial_groups <- add_unassigned_elements(target_groups, condition_blocked[select], N = N, K = K)
+    # ensure that previous conditions are still as before:
+    stopifnot(all(condition_blocked[select] == initial_groups, na.rm = TRUE))
+    # Now do anticlustering with restriction that previously assigned conditions must not change.
+    # Define who can be exchanged during anticlustering optimization (goes into the `categories` argument)
+    exchange_partners <- get_blocking_exchange_parters(condition_blocked, select)
+    condition_blocked[select] <- anticlustering(
+      input, 
+      K = initial_groups, 
+      objective = objective,
+      categories = cbind(exchange_partners, categories[select])
+    )
+    # ensure that previous conditions are still as before:
+    stopifnot(all(condition_blocked[blocks < i] == previous_groups[blocks < i], na.rm = TRUE)) 
+  }
+  
+  # verify that output has correct structure: 
+  tab <- table(blocks, condition_blocked)
+  stopifnot(all(abs(tab[ ,1] - tab[, 2]) <= 1)) # this test only works if equal group were requested, which are currently allowed only
+  condition_blocked
+}
+
+## Helper functions for anticlustering with memory for previous blocks:
+# Anticlustering with blocking, but with memory for assignment in previous blocks: 
+restricted_initialization <- function(groups_fixed, K, blocksize) {
+  groups_fixed[is.na(groups_fixed)] <- sample(rep_len(1:K, length.out = sum(is.na(groups_fixed))))
+  groups_fixed
+}
+
+# Create input for categories argument in anticlustering(), that ensures that previously blocked subjects
+# are not changed (only used to compute the anticlustering objective)
+get_blocking_exchange_parters <- function(condition_blocked, select) {
+  exchange_partners <- rep(1, sum(select))
+  is_already_assigned <- !is.na(condition_blocked[select])
+  groups_already_assigned <- condition_blocked[select][is_already_assigned]
+  exchange_partners[is_already_assigned] <- sample(sum(is_already_assigned), size = sum(is_already_assigned)) + 1
+  exchange_partners
+}